+++ packages/eas-cli/src/graphql/generated.ts
@@ -3719,39 +3719,6 @@ export type EditUpdateBranchMutation = (
   ) }
 );
 
-export type ViewBranchQueryVariables = Exact<{
-  appId: Scalars['String'];
-  name: Scalars['String'];
-  limit: Scalars['Int'];
-}>;
-
-
-export type ViewBranchQuery = (
-  { __typename?: 'RootQuery' }
-  & { app: (
-    { __typename?: 'AppQuery' }
-    & { byId: (
-      { __typename?: 'App' }
-      & Pick<App, 'id'>
-      & { updateBranchByName?: Maybe<(
-        { __typename?: 'UpdateBranch' }
-        & Pick<UpdateBranch, 'id' | 'name'>
-        & { updates: Array<(
-          { __typename?: 'Update' }
-          & Pick<Update, 'id' | 'group' | 'message' | 'createdAt' | 'runtimeVersion' | 'platform' | 'manifestFragment'>
-          & { actor?: Maybe<(
-            { __typename?: 'User' }
-            & Pick<User, 'username' | 'id'>
-          ) | (
-            { __typename?: 'Robot' }
-            & Pick<Robot, 'firstName' | 'id'>
-          )> }
-        )> }
-      )> }
-    ) }
-  ) }
-);
-
 export type CancelBuildMutationVariables = Exact<{
   buildId: Scalars['ID'];
 }>;
@@ -5399,6 +5366,39 @@ export type GetAllSubmissionsForAppQuery = (
         { __typename?: 'Submission' }
         & Pick<Submission, 'id'>
         & SubmissionFragment
+      )> }
+    ) }
+  ) }
+);
+
+export type ViewBranchUpdatesQueryVariables = Exact<{
+  appId: Scalars['String'];
+  name: Scalars['String'];
+  limit: Scalars['Int'];
+}>;
+
+
+export type ViewBranchUpdatesQuery = (
+  { __typename?: 'RootQuery' }
+  & { app: (
+    { __typename?: 'AppQuery' }
+    & { byId: (
+      { __typename?: 'App' }
+      & Pick<App, 'id'>
+      & { updateBranchByName?: Maybe<(
+        { __typename?: 'UpdateBranch' }
+        & Pick<UpdateBranch, 'id' | 'name'>
+        & { updates: Array<(
+          { __typename?: 'Update' }
+          & Pick<Update, 'id' | 'group' | 'message' | 'createdAt' | 'runtimeVersion' | 'platform' | 'manifestFragment'>
+          & { actor?: Maybe<(
+            { __typename?: 'User' }
+            & Pick<User, 'username' | 'id'>
+          ) | (
+            { __typename?: 'Robot' }
+            & Pick<Robot, 'firstName' | 'id'>
+          )> }
+        )> }
       )> }
     ) }
   ) }